Introduce Task that refreshes data constantly
It was getting rather annoying to restart the app every time data was
added to the database.

This change addresses that by adding a background task that retrieves
the newest data every ten seconds. It also preserves the selection of
the user for a better user experience.

This change has no side effects.

Not exactly keeping track of issues in Github, sorry.