Add progress bars to the right pane
It's a more visual representation of the nutrient progress and it gives
the UI a little bit more colour. But mostly to track progress towards
daily goals.

This change adds the Goals and GoalProgress User controls in order to
render the bars and introduces a struct where the daily goals are
stored.

There has been some code clean-up, too, hence the long file list.

No ticket for this.