Refactor code to use a Goal class
The GoalProgress class had a property for every type of goal and some
calculations. It made more sense to put that relevant info into a Goal
class.

This commit introduces the Goal class. This class takes the duplicated
properties and logic and puts it in its own class.

The SetGoals and GoalProgress classes could be deleted, since the
progress toward a goal became an intrinsic property of the goal itself.

Nope, still no tickets.