Improve reload functionality
It was a bit annoying that the list would be updated while you're
browsing through it.

This change adds items to the underlying collection and
doesn't replace the entire collection. Instead, a CollectionViewSource
is used to bind against.

No side effects.

Still no tickets for this hobby project to link to. :D